Last fix of all wrestler tests
