+++ tests/Feature/Admin/Wrestlers/CreateWrestlerSuccessConditionsTest.php
@@ -2,9 +2,7 @@
 
 namespace Tests\Feature\Admin\Wrestlers;
 
-use Carbon\Carbon;
 use Tests\TestCase;
-use App\Models\User;
 use App\Models\Wrestler;
 use Illuminate\Foundation\Testing\RefreshDatabase;
 
@@ -38,10 +36,9 @@ class CreateWrestlerSuccessConditionsTest extends TestCase
     /** @test */
     public function an_administrator_can_view_the_form_for_creating_a_wrestler()
     {
-        $user = factory(User::class)->states('administrator')->create();
+        $this->actAs('administrator');
 
-        $response = $this->actingAs($user)
-                        ->get(route('wrestlers.create'));
+        $response = $this->get(route('wrestlers.create'));
 
         $response->assertViewIs('wrestlers.create');
         $response->assertViewHas('wrestler', new Wrestler);
@@ -50,10 +47,9 @@ class CreateWrestlerSuccessConditionsTest extends TestCase
     /** @test */
     public function an_administrator_can_create_a_wrestler()
     {
-        $user = factory(User::class)->states('administrator')->create();
+        $this->actAs('administrator');
 
-        $response = $this->actingAs($user)
-                        ->from(route('wrestlers.create'))
+        $response = $this->from(route('wrestlers.create'))
                         ->post(route('wrestlers.store'), $this->validParams());
 
         $response->assertRedirect(route('wrestlers.index'));
@@ -64,57 +60,6 @@ class CreateWrestlerSuccessConditionsTest extends TestCase
             $this->assertEquals('Laraville, FL', $wrestler->hometown);
             $this->assertEquals('The Finisher', $wrestler->signature_move);
             $this->assertEquals(today()->toDateString(), $wrestler->hired_at->toDateString());
-        });
-    }
-
-    /** @test */
-    public function a_wrestler_hired_today_or_before_is_bookable()
-    {
-        $user = factory(User::class)->states('administrator')->create();
-
-        $this->actingAs($user)
-            ->from(route('wrestlers.create'))
-            ->post(route('wrestlers.store'), $this->validParams([
-                'hired_at' => today()->toDateTimeString()
-            ]));
-
-        tap(Wrestler::first(), function ($wrestler) {
-            $this->assertTrue($wrestler->is_bookable);
-        });
-    }
-
-    /** @test */
-    public function a_wrestler_hired_after_today_is_not_bookable()
-    {
-        $user = factory(User::class)->states('administrator')->create();
-
-        $this->actingAs($user)
-            ->from(route('wrestlers.create'))
-            ->post(route('wrestlers.store'), $this->validParams([
-                'hired_at' => Carbon::tomorrow()->toDateTimeString()
-            ]));
-
-        tap(Wrestler::first(), function ($wrestler) {
-            $this->assertFalse($wrestler->is_bookable);
-            $this->assertFalse($wrestler->is_hired);
-        });
-    }
-
-    /** @test */
-    public function a_wrestler_signature_move_is_optional()
-    {
-        $user = factory(User::class)->states('administrator')->create();
-
-        $response = $this->actingAs($user)
-                        ->from(route('wrestlers.create'))
-                        ->post(route('wrestlers.store'), $this->validParams([
-                            'signature_move' => '',
-                        ]));
-
-        $response->assertSessionDoesntHaveErrors('signature_move');
-        $response->assertRedirect(route('wrestlers.index'));
-        tap(Wrestler::first(), function ($wrestler) {
-            $this->assertNull($wrestler->signature_move);
         });
     }
 }
