+++ tests/Feature/User/Wrestlers/ViewWrestlerBioPageSuccessConditionsTest.php
@@ -2,7 +2,6 @@
 
 namespace Tests\Feature\User\Wrestlers;
 
-use App\Models\User;
 use App\Models\Wrestler;
 use Tests\TestCase;
 use Illuminate\Foundation\Testing\RefreshDatabase;
@@ -15,7 +14,6 @@ class ViewWrestlerBioPageSuccessConditionsTest extends TestCase
     public function a_basic_user_can_view_their_wrestler_profile()
     {
         $signedInUser = $this->actAs('basic-user');
-
         $wrestler = factory(Wrestler::class)->create(['user_id' => $signedInUser->id]);
 
         $response = $this->get(route('wrestlers.show', ['wrestler' => $wrestler]));
